🚚 chore: create app folder for better repository architecture
