+++ src/app/leagues/types/leagues.ts
