+++ src/main.ts
@@ -5,7 +5,7 @@ import { ConfigService } from '@nestjs/config';
 
 import helmet from 'helmet';
 
-import { AppModule } from './app.module';
+import { AppModule } from './app/app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create<NestExpressApplication>(AppModule);
